pyproject.toml
[project]
name = "sphene"
version = "0.1.0"
description = "OpenAI-powered Discord bot with XIVAPI integration"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai",
    "discord.py>=2.0.0",
    "python-dotenv",
    "google-cloud-firestore",
    "requests",
    "httpx",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.1",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "mypy",
    "types-requests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = "--verbose --cov=. --cov-report=term --cov-report=html --no-cov-on-fail"

[tool.mypy]
exclude = ["scripts/"]

[[tool.mypy.overrides]]
module = ["google.*"]
ignore_missing_imports = true
follow_untyped_imports = true

[tool.coverage.run]
omit = ["tests/*", "setup.py", "scripts/*"]
